Move ErrorMessages component

# Conflicts:
#	src/components/MainHeader/MainHeader.tsx
